# feat: improve prettier performance by running Promise.all()