*Issue #, if available:*
Fixes: https://github.com/aws/aws-sdk-js-v3/issues/1947

*Description of changes:*
improves prettier performance by running prettier for each directory in a `Promise.all()`

Metrics to generate all clients:

<details>
<summary>Existing setup takes ~360 seconds</summary>

```console
$ time yarn generate-clients
...
...
Done in 359.39s.
433.54s user 13.15s system 124% cpu 5:59.52 total
```

</details>

<details>
<summary>Disabling prettier takes ~17 seconds</summary>

```console
$ time yarn generate-clients
...
...
Done in 17.10s.
6.29s user 1.55s system 45% cpu 17.262 total
```

</details>

<details>
<summary>Using Promise.all() takes ~150 seconds</summary>

```console
$ time yarn generate-clients
...
...
Done in 151.02s.
1857.47s user 62.18s system 1270% cpu 2:31.15 total
```

</details>

This improves time taken to generate all clients by ~58% on a 16 core CPU.

<details>
<summary>System Information</summary>

* Operating System: Amazon Linux 2 x86_64
* Host type: c5.4xlarge (16 Virtual CPUs, 32.0 GiB Memory, 160 GiB SSD (EBS) Storage)
* Node.js version: v14.15.4

</details>

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.