@trivikr 2021-01-22
Superceded by https://github.com/aws/aws-sdk-js-v3/pull/1949